Bubble Sheets From CSV File!!!
You can now create bubble sheets from a CSV file with each competitors' 9-digit ID and name. On the page there is a link to an example file for how the CSV should be formatted.

Note: There currently is no link in the nav-bar. This will be available in the next big release (coming soon, :>).

-Robert C.